cors: frontend - added Readme
