+++ frontend/README.md
@@ -1,0 +1,26 @@
+# M-Uwezo App
+
+## Backend Setup
+
+The first thing to do is to clone the repository:
+
+```sh
+$ git clone https://github.com/BuildForSDG/team-159.git
+$ cd team-159/frontend
+```
+
+TO build the app
+
+```sh
+$ npm install
+$ npm run build
+```
+
+To start the application
+
+```sh
+$ npm run start
+```
+
+And navigate to `http://127.0.0.1:3000/`.
+
