--- frontend/src/main.js
@@ -1,12 +1,0 @@
-import React from 'react'
-import { BrowserRouter } from 'react-router-dom'
-import { hydrate } from 'react-dom'
-
-import App from './App'
-
-hydrate(
-  <BrowserRouter>
-    <App />
-  </BrowserRouter>,
-  document.getElementById('app')
-)
